refactor(tests): extract docker test runner to dedicated module
Move all docker test runner logic from docker_tests_main.rs to a new
docker_tests/runner.rs module. This simplifies the main entry point to
a thin wrapper and improves maintainability.

Key changes:
- Create DockerTestRunner struct with RAII container management
- Use Vec<Box<dyn Any>> for type-erased container keep-alive
- Replace if/else chains with match statements for test modes
- Feature-gate setup methods with explicit #[cfg] attributes
- Fix UTXO metadata to store ports consistently (8000/8001)
- Fix _MM2_TEST_CONF to allow metadata reuse mode
- Reduce docker_tests_main.rs from 823 to 46 lines
