+++ mm2src/mm2_main/tests/docker_tests/mod.rs
@@ -1,6 +1,7 @@
 #![allow(static_mut_refs)]
 
 pub mod docker_env_metadata;
+pub mod runner;
 
 // Helpers are used by all docker tests, and also by some sepolia tests
 #[cfg(any(
